Add music player component and song list item for enhanced music playback experience
